requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
Django==5.2.1
mysql-connector-python==9.3.0
mysqlclient==2.2.7
pycparser==2.22
sqlparse==0.5.3
